Change sensor reg_addr to uint16_t

+++ src/omv/lepton.c
@@ -115,7 +115,7 @@ static int sleep(sensor_t *sensor, int enable)
     return 0;
 }
 
-static int read_reg(sensor_t *sensor, uint8_t reg_addr)
+static int read_reg(sensor_t *sensor, uint16_t reg_addr)
 {
     uint16_t reg_data;
     if (cambus_readw2(sensor->slv_addr, reg_addr, &reg_data)) {
@@ -124,7 +124,7 @@ static int read_reg(sensor_t *sensor, uint8_t reg_addr)
     return reg_data;
 }
 
-static int write_reg(sensor_t *sensor, uint8_t reg_addr, uint16_t reg_data)
+static int write_reg(sensor_t *sensor, uint16_t reg_addr, uint16_t reg_data)
 {
     return cambus_writew2(sensor->slv_addr, reg_addr, reg_data);
 }
